# backend/docs/api-spec.yaml
openapi: 3.0.3
info:
  title: Sports Management API
  description: |
    Comprehensive API for managing sports leagues, referee assignments, games, teams, and budgets.
    
    ## Authentication
    All endpoints (except `/health` and `/auth/login`) require Bearer authentication.
    Include the JWT token in the Authorization header: `Authorization: Bearer <token>`
    
    ## Rate Limiting
    - Login endpoint: 5 requests per minute
    - Registration endpoint: 3 requests per hour
    - Password reset: 3 requests per hour
    - Assignment operations: 100 requests per minute
    - General endpoints: 1000 requests per hour
    
    ## Error Handling
    All error responses follow a consistent format:
    ```json
    {
      "error": "Error message",
      "message": "Detailed description",
      "code": "ERROR_CODE"
    }
    ```
    
    ## Data Models
    The API follows a hierarchical structure:
    - Organizations contain Leagues
    - Leagues contain Teams
    - Games reference Teams and are assigned Referees
    - Budgets track financial aspects
    
  version: 1.0.0
  contact:
    name: Sports Management API Support
    email: support@sportsmanagement.app
  license:
    name: MIT
servers:
  - url: http://localhost:3001/api
    description: Development server
  - url: https://api.sportsmanagement.app
    description: Production server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /auth/login
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Unique user identifier
        email:
          type: string
          format: email
          description: User's email address
        role:
          type: string
          enum: [admin, referee, manager]
          description: User's primary role
        roles:
          type: array
          items:
            type: string
          description: Array of user roles (new role system)
        name:
          type: string
          description: User's full name
        phone:
          type: string
          description: User's phone number
        location:
          type: string
          description: User's location/address
        postal_code:
          type: string
          maxLength: 10
          description: User's postal code for distance calculations
        max_distance:
          type: integer
          minimum: 1
          maximum: 200
          default: 25
          description: Maximum travel distance in km (for referees)
        is_available:
          type: boolean
          description: Whether the user is available for assignments
        wage_per_game:
          type: number
          minimum: 0
          description: User's wage per game (for referees)
        referee_level_id:
          type: string
          format: uuid
          description: Reference to referee level
        years_experience:
          type: integer
          minimum: 0
          maximum: 50
          description: Years of experience
        games_refereed_season:
          type: integer
          minimum: 0
          description: Number of games refereed this season
        evaluation_score:
          type: number
          minimum: 0
          maximum: 10
          description: Average evaluation score
        white_whistle:
          type: boolean
          description: Whether referee has white whistle certification
        notes:
          type: string
          description: Additional notes about the user
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
        - id
        - email
        - role
...
    PaginatedResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
        pagination:
          type: object
          properties:
            page:
              type: integer
            limit:
              type: integer
            total:
              type: integer
            totalPages:
              type: integer
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Error message
        message:
          type: string
          description: Detailed error description
        code:
          type: string
          description: Error code for programmatic handling
...
security:
  - bearerAuth: []
paths:
...
  /referees:
    get:
      tags:
        - Referees
      summary: Get all referees
      description: Retrieve all referees with optional filtering
      parameters:
        - name: level
          in: query
          description: Filter by referee level
          schema:
            type: string
        - name: postal_code
          in: query
          description: Filter by postal code
          schema:
            type: string
        - name: is_available
          in: query
          description: Filter by availability status
          schema:
            type: boolean
        - name: white_whistle
          in: query
          description: Filter by white whistle certification
          schema:
            type: boolean
        - name: search
          in: query
          description: Search by name or email
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
      responses:
        '200':
          description: Referees retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedResponse'
  /referees/{id}:
    get:
      tags:
        - Referees
      summary: Get specific referee
      description: Retrieve detailed information about a specific referee
      parameters:
        - name: id
          in: path
          required: true
          description: Referee ID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Referee retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: '#/components/schemas/User'
        '404':
          description: Referee not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # Leagues
...